+++ lib/src/resources/app_assets.dart
@@ -2,5 +2,5 @@ final class AppAssets {
   const AppAssets._();
 
   static const imagesLogo = 'assets/images/logo.png';
-  static const fakeAvatarImage = 'assets/images/fake_avatar.jpg';
+  static const imagesFakeAvatar = 'assets/images/fake_avatar.jpg';
 }
